move row_filter code and Extract page filter code into its own module
