+++ datafusion/core/src/physical_plan/file_format/mod.rs
@@ -25,7 +25,6 @@ mod delimited_stream;
 mod file_stream;
 mod json;
 mod parquet;
-mod row_filter;
 
 pub(crate) use self::csv::plan_to_csv;
 pub use self::csv::CsvExec;
